+++ src/Marketing/Enums/InventoryCriterionEnum.php
@@ -12,7 +12,7 @@ namespace DTS\eBaySDK\Marketing\Enums;
 
 class InventoryCriterionEnum
 {
-    const C_INVALID = 'INVALID';
+    const C_INVENTORY_ANY = 'INVENTORY_ANY';
     const C_INVENTORY_BY_RULE = 'INVENTORY_BY_RULE';
     const C_INVENTORY_BY_VALUE = 'INVENTORY_BY_VALUE';
 }